[Refactor/#48]: 댓글 관련 hook 통합
