+++ .eslintrc.json
@@ -33,6 +33,7 @@
     "no-plusplus": "off",
     "no-nested-ternary": "warn",
     "class-methods-use-this": "off",
+    "@typescript-eslint/ban-types": "off",
     "@typescript-eslint/lines-between-class-members": "off",
     "import/no-extraneous-dependencies": "off",
     "react/jsx-no-useless-fragment": "off",
